Backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.4
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.0
elastic-transport==8.15.0
elasticsearch==8.15.1
fastapi==0.115.0
fastavro==1.9.7
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.0
idna==3.10
jiter==0.5.0
jmespath==1.0.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-cohere==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-elasticsearch==0.3.0
langchain-experimental==0.3.0
langchain-text-splitters==0.3.0
langsmith==0.1.125
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.47.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
passlib==1.7.4
pdf2image==1.17.0
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
simsimd==5.4.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tabulate==0.9.0
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.11.1
